requirements.txt
torch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1
anthropic==0.75.0
openai==2.11.0
google-genai==1.56.0
google-generativeai==0.8.6
litellm==1.80.5
mistralai==1.10.0
transformers==4.57.3
accelerate==1.12.0
peft==0.18.0
sentence-transformers==5.2.0
diffusers==0.36.0
timm==1.0.22
vllm==0.7.3
xformers==0.0.28.post3
einops==0.8.1
datasets==4.4.1
huggingface-hub==0.36.0
safetensors==0.7.0
tokenizers==0.22.1
sentencepiece==0.2.1
tiktoken==0.12.0
numpy==1.26.4
pandas==2.3.3
scipy==1.16.3
scikit-learn==1.8.0
faiss-cpu==1.13.2
wandb==0.23.1
ray==2.40.0
fastapi==0.128.0
uvicorn==0.40.0
pillow==12.0.0
opencv-python-headless==4.11.0.86
librosa==0.11.0
tqdm==4.67.1
pyyaml==6.0.3
python-dotenv==1.2.1
requests==2.32.5
httpx==0.28.1
jinja2==3.1.6
rich==14.2.0
pytest==9.0.2
decord==0.6.0
onnxruntime==1.23.2
omegaconf==2.3.0